Impove library initialization and refactor files grouped by control
