+++ src/HorusStudio.Maui.MaterialDesignControls/Controls/Button/Button.Windows.cs
