+++ src/HorusStudio.Maui.MaterialDesignControls/MaterialDesignControlsBuilder.cs
@@ -1,15 +1,74 @@
-﻿namespace HorusStudio.Maui.MaterialDesignControls
+﻿using HorusStudio.Maui.MaterialDesignControls.Utils;
+using Microsoft.Maui.LifecycleEvents;
+
+namespace HorusStudio.Maui.MaterialDesignControls
 {
     public static class MaterialDesignControlsBuilder
     {
         public static MauiAppBuilder ConfigureMaterialDesignControls(this MauiAppBuilder builder)
         {
-            builder.ConfigureMauiHandlers(handlers =>
-            {
-                handlers.AddHandler(typeof(CustomButton), typeof(Handlers.CustomButtonHandler));
-            });
-
+            builder
+                .ConfigureMauiHandlers(ConfigureHandlers)
+                .ConfigureLifecycleEvents(ConfigureLifeCycleEvents);
+            
             return builder;
+        }
+
+        private static void ConfigureHandlers(IMauiHandlersCollection handlers)
+        {
+            handlers.AddHandler(typeof(CustomButton), typeof(CustomButtonHandler));
+        }
+
+        private static void ConfigureLifeCycleEvents(ILifecycleBuilder appLifeCycle)
+        {
+#if ANDROID
+            appLifeCycle.AddAndroid(android => android
+                .OnApplicationCreate(_ =>
+                {
+                    InitializeComponents();
+                }));
+#elif IOS || MACCATALYST
+            appLifeCycle.AddiOS(ios => ios
+                .FinishedLaunching((_, _) =>
+                {
+                    InitializeComponents();
+                    return true;
+                }));
+#elif WINDOWS
+            appLifeCycle.AddWindows(windows => windows
+                .OnLaunched((_, _) =>
+                {
+                    InitializeComponents();
+                }
+#endif
+        }
+
+        private static void InitializeComponents()
+        {
+            if (Application.Current == null)
+            {
+                Logger.Debug("Error initializing Material Design Controls: MAUI Application is null.");
+                return;
+            }
+            RegisterDefaultStyles(Application.Current);
+        }
+
+        private static void RegisterDefaultStyles(Application application)
+        {
+            // Button
+            application.Resources.AddStyles(MaterialButton.GetStyles());
+            // Icon Button
+            application.Resources.AddStyles(MaterialIconButton.GetStyles());
+        }
+
+        private static void AddStyles(this ResourceDictionary resources, IEnumerable<Style> styles)
+        {
+            if (styles == null) return;
+
+            foreach (var style in styles)
+            {
+                resources.Add(style);
+            }
         }
     }
 }
